requirements.txt
attrs==24.2.0
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.3.2
exceptiongroup==1.2.2
flatbuffers==24.3.25
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.10
numpy==2.0.2
oauthlib==3.2.2
openmeteo_requests==1.3.0
openmeteo_sdk==1.15.0
pandas==2.2.3
platformdirs==4.3.6
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
retry-requests==2.0.0
rsa==4.9
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.2.3
